old/site/requirements.txt
flask==3.0.2
pandas==2.2.0
python-dotenv==1.0.1
pyzk==0.9 